requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.22
python-doctr[torch]==0.8.1
pillow==10.3.0
pydantic==2.6.1
pydantic-settings==2.1.0
requests==2.31.0
pytest==8.0.0
httpx==0.26.0
pymupdf==1.24.0
opencv-python==4.9.0.80
numpy>=1.24.0
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
jinja2==3.1.3
aiofiles==23.2.1
